+++ javascript/modules/client/test/integration/client-erc20.test.ts
@@ -156,7 +156,7 @@ describe("Client", () => {
       const client = new ClientErc20(context)
 
       const proposalParams: ICreateProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         metadata: {
           title: 'Best Proposal',
           summary: 'this is the sumnary',
@@ -201,7 +201,7 @@ describe("Client", () => {
         })
 
       const proposalParams: ICreateProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         metadata: {
           title: 'Best Proposal',
           summary: 'this is the sumnary',
@@ -247,7 +247,7 @@ describe("Client", () => {
       const client = new ClientErc20(context)
 
       const voteParams: IVoteProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         proposalId: '0x1234567890123456789012345678901234567890',
         vote: VoteValues.YES
       }
@@ -266,7 +266,7 @@ describe("Client", () => {
       const client = new ClientErc20(context)
 
       const voteParams: IVoteProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         proposalId: '0x1234567890123456789012345678901234567890',
         vote: VoteValues.YES
       }
@@ -298,7 +298,7 @@ describe("Client", () => {
 
 
       const executeParams: IExecuteProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         proposalId: '0x1234567890123456789012345678901234567890',
       }
       const estimation = await client.estimation.executeProposal(executeParams)
@@ -316,7 +316,7 @@ describe("Client", () => {
       const client = new ClientErc20(context)
 
       const executeParams: IExecuteProposalParams = {
-        pluginInstanceAddress: "0x123456789012345678901234567890123456789012",
+        pluginAddress: "0x123456789012345678901234567890123456789012",
         proposalId: '0x1234567890123456789012345678901234567890',
       }